Added debugging and configuration to child process.
- Added basic options to spawn function. It is not clear what is
required for a language client. After adding this (and the on error code
noted below) the startup time for the package jumped from milliseconds
to two seconds. Will need to see why that is.
- In an attempt to get some information out of the ide I added an on
error function with the goal to output any errors occured. None are
showing up at this time.